refactor(game): extract run lifecycle to game/lifecycle.ts; drop dead addTestHooks (R2.6)
Final R2 step. Move reset/restart/camera-toggle out of snowglider.ts into
`src/game/lifecycle.ts` as `createLifecycle(deps)`, returning `resetSnowman` /
`restartGame` / `toggleCameraView` plus `initLifecycleUI()` (which wires the reset
button, the camera-toggle button — created+appended here — and the restart button).
The coordinator re-publishes the three hooks on `window` and calls `initLifecycleUI()`.

Deletes the dead local `addTestHooks` shim (never called by bare name; all live sites
call `Snowman.addTestHooks`). It was a stale, incomplete duplicate that only installed
`forceTreeCollision`; the real browser hooks come from `Snowman.addTestHooks`, which
the coordinator still calls eagerly. `publishGameGlobals()`, the `isTestMode` flag,
and the test-mode auto-start stay in the coordinator.

Mechanical, behavior-preserving: every `window.*` hook and the publishGameGlobals
proxy set are unchanged. Dropped now-unused coordinator imports CourseModule /
EffectsModule.

After R2, snowglider.ts is a ~380-line coordinator (was 1380).

Verification: lint (0 errors) / typecheck / test (contract guard 44/44) / build
green; puppeteer 87/87 across 7/7 suites (controls/camera drive reset/restart/toggle).
