+++ src/snowglider.ts
@@ -30,14 +30,13 @@ import * as THREE from 'three';
 import { Controls } from './controls.js';
 import { Snow } from './snow.js';
 import { Snowman } from './snowman.js';
-import { CourseModule } from './course.js';
-import { EffectsModule } from './effects.js';
 import { AudioModule } from './audio.js';
 import { Physics } from './physics.js';
 import { initializeGameStats, initializeControlsToggle, updateTimerDisplay } from './ui/hud.js';
 import { readStoredBestTime, createShowGameOver } from './ui/result-overlay.js';
 import { setupScene } from './game/scene-setup.js';
 import { createMainLoop } from './game/main-loop.js';
+import { createLifecycle } from './game/lifecycle.js';
 
 // Get keyboard controls from the Controls module
 Controls.setupControls();
@@ -92,73 +91,6 @@ bestTimeDisplay.style.fontSize = '20px';
 bestTimeDisplay.style.marginBottom = '20px';
 gameOverOverlay.insertBefore(bestTimeDisplay, restartButton);
 
-function resetSnowman() {
-  // Reset the snowman + player physics state (position, velocity, camera, and the
-  // air/auto-turn scalars) to the start of a run.
-  Physics.resetPlayer(player, snowman, Snow.getTerrainHeight, cameraManager);
-
-  // Reset avalanche system
-  const avalanche = state.avalanche;
-  if (avalanche) {
-    avalanche.reset();
-    state.avalancheTriggered = false;
-    state.lastAvalancheZ = pos.z; // Reset to starting position
-  }
-  
-  // Reset keyboard controls
-  Controls.resetControls();
-  
-  state.startTime = performance.now(); // Reset the timer when starting a new run
-  updateTimerDisplay(state.gameActive, state.startTime, state.bestTime);
-  
-  // Reset course (gates/splits/ghost) and effects (avalanche UI, FOV, shake) for the new run
-  if (CourseModule) CourseModule.reset();
-  if (EffectsModule) EffectsModule.reset();
-  
-  // Track game reset in Analytics if available
-  try {
-    // Only try to use analytics when properly initialized with modular SDK
-    if (window.firebaseModules && typeof window.firebaseModules.logEvent === 'function' && window.location.protocol !== 'file:') {
-      // Using the direct logEvent function
-      window.firebaseModules.logEvent('game_reset');
-    }
-  } catch (e) {
-    console.log("Analytics tracking skipped:", (e as Error).message);
-  }
-}
-// Make resetSnowman accessible globally for touch handler
-window.resetSnowman = resetSnowman;
-
-// Initialize controls but don't reset the snowman yet
-document.getElementById('resetBtn')!.addEventListener('click', resetSnowman);
-
-// Add camera toggle button
-const cameraToggleBtn = document.createElement('button');
-cameraToggleBtn.id = 'cameraToggleBtn';
-cameraToggleBtn.textContent = 'Toggle Chase View';
-cameraToggleBtn.style.position = 'absolute';
-cameraToggleBtn.style.bottom = '20px';
-cameraToggleBtn.style.left = '170px'; // Position it next to reset button
-cameraToggleBtn.style.padding = '15px 20px';
-cameraToggleBtn.style.border = 'none';
-cameraToggleBtn.style.borderRadius = '8px';
-cameraToggleBtn.style.backgroundColor = '#4a69bd'; // Different color from reset button
-cameraToggleBtn.style.color = 'white';
-cameraToggleBtn.style.cursor = 'pointer';
-cameraToggleBtn.style.fontSize = '16px';
-cameraToggleBtn.style.setProperty('-webkit-tap-highlight-color', 'rgba(255, 255, 255, 0.5)');
-cameraToggleBtn.style.touchAction = 'manipulation'; // Removes delay on mobile devices
-cameraToggleBtn.style.userSelect = 'none';
-
-// Add both click and touchend events to ensure cross-platform compatibility
-cameraToggleBtn.addEventListener('click', toggleCameraView);
-cameraToggleBtn.addEventListener('touchend', function(event) {
-  event.preventDefault();
-  toggleCameraView();
-}, { passive: false });
-
-document.body.appendChild(cameraToggleBtn);
-
 // --- Initial Camera Setup ---
 // Initialize camera with the snowman's position and rotation
 cameraManager.initialize(
@@ -197,131 +129,30 @@ const { updateSnowman, updateCamera, startLoop, handleResize } = createMainLoop(
 });
 window.addEventListener('resize', handleResize);
 
-function restartGame() {
-  gameOverOverlay.style.display = 'none';
-  state.gameActive = true;
-
-  // Clear the finish result panel from the previous run, if present.
-  const oldResult = document.getElementById('courseResult');
-  if (oldResult && oldResult.parentNode) oldResult.parentNode.removeChild(oldResult);
-  
-  // Add game-active class to body for styling
-  document.body.classList.add('game-active');
-  
-  // Show and reset game stats
-  const gameStatsContainer = document.getElementById('gameStatsContainer');
-  if (gameStatsContainer) {
-    gameStatsContainer.classList.remove('collapsed');
-    const toggleBtn = document.getElementById('toggleStats');
-    if (toggleBtn) {
-      toggleBtn.textContent = '▲';
-    }
-    
-    // Reset colors for best time display
-    const bestTimeElement = document.getElementById('bestTimeValue');
-    if (bestTimeElement) {
-      bestTimeElement.style.color = ''; // Reset to default color
-    }
-  }
-  
-  resetSnowman();
-  
-  // Initialize camera with the snowman's position and rotation
-  cameraManager.initialize(snowman.position, snowman.rotation);
-  
-  // TODO: AUDIO DISABLED - Resume audio (will be no-op if disabled)
-  if (AudioModule) {
-    AudioModule.enableSound(true);
-  }
-  
-  // Reset animation if it was stopped
-  if (!state.animationRunning) {
-    state.animationRunning = true;
-    startLoop();
-  }
-}
-// Make restartGame accessible globally for touch handler
+// --- Run lifecycle (see game/lifecycle.ts): reset / restart / camera toggle ---
+// Built after the loop (restartGame uses startLoop). Re-publish the three hooks the
+// touch handlers + controls.js drive by bare name, then wire the DOM controls.
+const { resetSnowman, restartGame, toggleCameraView, initLifecycleUI } = createLifecycle({
+  state,
+  cameraManager,
+  snowman,
+  gameOverOverlay,
+  restartButton,
+  player,
+  startLoop,
+});
+// Make reset/restart/toggle accessible globally for touch + keyboard handlers.
+window.resetSnowman = resetSnowman;
 window.restartGame = restartGame;
+window.toggleCameraView = toggleCameraView;
+initLifecycleUI();
 
 // Make showGameOver accessible globally for test hooks
 window.showGameOver = showGameOver;
-
-// Toggle between first-person and third-person camera views
-function toggleCameraView() {
-  // Call the camera manager's toggle method
-  const newMode = cameraManager.toggleCameraMode();
-  
-  // Reset camera initialization with current snowman position and rotation
-  cameraManager.initialize(snowman.position, snowman.rotation);
-  
-  // Update the camera mode text in the controls info
-  const viewControlItem = document.querySelector('#controlsContent .control-item:last-child');
-  if (viewControlItem) {
-    const keyBadge = viewControlItem.querySelector('.key-badge');
-    const textSpan = viewControlItem.querySelector('span:last-child');
-    
-    if (keyBadge && textSpan) {
-      keyBadge.textContent = 'V';
-      textSpan.textContent = `Toggle ${newMode === 'thirdPerson' ? 'Normal' : 'Chase'} View`;
-    }
-  }
-  
-  // Update the toggle button text
-  const cameraToggleBtn = document.getElementById('cameraToggleBtn');
-  if (cameraToggleBtn) {
-    cameraToggleBtn.textContent = `Toggle ${newMode === 'thirdPerson' ? 'Normal' : 'Chase'} View`;
-  }
-  
-  // Return the new mode (useful for tests)
-  return newMode;
-}
-
-// Make toggleCameraView accessible globally for the keyboard handler in controls.js
-window.toggleCameraView = toggleCameraView;
-
-// Add test hook functions for tree collision testing
-function addTestHooks(
-  pos: { x: number; y: number; z: number },
-  showGameOver: (reason: string) => void,
-  getTerrainHeight: TerrainHeightFn
-) {
-  console.log("Snowman.addTestHooks called - setting up test hooks");
-  
-  if (!window.testHooks) {
-    window.testHooks = {};
-  }
-  
-  // Add a force collision function that can be called by tests
-  window.testHooks.forceTreeCollision = function() {
-    console.log("TEST: forceTreeCollision hook called");
-    
-    // Create a direct function reference to ensure it works
-    const directShowGameOver = window.showGameOver || showGameOver;
-    
-    // Call the function directly to ensure it works
-    console.log("TEST: Forcing tree collision (direct call)");
-    try {
-      directShowGameOver("BANG!!! You hit a tree!");
-      console.log("TEST: Successfully called showGameOver");
-    } catch (error) {
-      console.error("TEST ERROR: Failed to call showGameOver:", error);
-      // As a fallback, just simulate the collision for the test
-      window.testCollisionDetected = true;
-    }
-    
-    return true;
-  };
-  
-  // Other test hooks and functionality...
-}
-
 // Initialize test hooks explicitly to ensure they're available immediately
 // This is important for browser tests that run soon after page load
 console.log("Initializing test hooks on startup");
 Snowman.addTestHooks(pos, showGameOver, Snow.getTerrainHeight);
-
-// Add event listener to restart button
-restartButton.addEventListener('click', restartGame);
 
 // Initialize controls toggle when DOM is loaded
 document.addEventListener('DOMContentLoaded', function() {
